+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-add-ons/dynamic-form-add-ons.module.ts
@@ -1,0 +1,18 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { DynamicFormElementModule } from '@dynamic-forms/core';
+import { BsDynamicFormAddOnsComponent } from './dynamic-form-add-ons.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    DynamicFormElementModule,
+  ],
+  declarations: [
+    BsDynamicFormAddOnsComponent,
+  ],
+  exports: [
+    BsDynamicFormAddOnsComponent,
+  ],
+})
+export class BsDynamicFormAddOnsModule {}