+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-combobox/dynamic-form-combobox.module.ts
@@ -3,6 +3,8 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormConfigModule, DynamicFormInputType } from '@dynamic-forms/core';
 import { bsDynamicFormLibrary } from '../../dynamic-form-library/dynamic-form-library';
+import { BsDynamicFormAddOnsModule } from '../dynamic-form-add-ons/dynamic-form-add-ons.module';
+
 import { BsDynamicFormComboboxComponent } from './dynamic-form-combobox.component';
 
 export const bsDynamicFormComboboxType: DynamicFormInputType = {
@@ -17,6 +19,7 @@ export const bsDynamicFormComboboxType: DynamicFormInputType = {
     CommonModule,
     ReactiveFormsModule,
     DynamicFormConfigModule.withInput(bsDynamicFormComboboxType),
+    BsDynamicFormAddOnsModule,
   ],
   declarations: [
     BsDynamicFormComboboxComponent,