+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-datepicker/dynamic-form-datepicker.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormConfigModule, DynamicFormInputType } from '@dynamic-forms/core';
 import { bsDynamicFormLibrary } from '../../dynamic-form-library/dynamic-form-library';
+import { BsDynamicFormAddOnsModule } from '../dynamic-form-add-ons/dynamic-form-add-ons.module';
 import { BsDynamicFormDatepickerComponent } from './dynamic-form-datepicker.component';
 
 export const bsDynamicFormDatepickerType: DynamicFormInputType = {
@@ -17,6 +18,7 @@ export const bsDynamicFormDatepickerType: DynamicFormInputType = {
     CommonModule,
     ReactiveFormsModule,
     DynamicFormConfigModule.withInput(bsDynamicFormDatepickerType),
+    BsDynamicFormAddOnsModule,
   ],
   declarations: [
     BsDynamicFormDatepickerComponent,