+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-numberbox/dynamic-form-numberbox.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormConfigModule, DynamicFormInputType } from '@dynamic-forms/core';
 import { bsDynamicFormLibrary } from '../../dynamic-form-library/dynamic-form-library';
+import { BsDynamicFormAddOnsModule } from '../dynamic-form-add-ons/dynamic-form-add-ons.module';
 import { BsDynamicFormNumberboxComponent } from './dynamic-form-numberbox.component';
 
 export const bsDynamicFormNumberboxType: DynamicFormInputType = {
@@ -17,6 +18,7 @@ export const bsDynamicFormNumberboxType: DynamicFormInputType = {
     CommonModule,
     ReactiveFormsModule,
     DynamicFormConfigModule.withInput(bsDynamicFormNumberboxType),
+    BsDynamicFormAddOnsModule,
   ],
   declarations: [
     BsDynamicFormNumberboxComponent,