+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-select/dynamic-form-select.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormConfigModule, DynamicFormInputType } from '@dynamic-forms/core';
 import { bsDynamicFormLibrary } from '../../dynamic-form-library/dynamic-form-library';
+import { BsDynamicFormAddOnsModule } from '../dynamic-form-add-ons/dynamic-form-add-ons.module';
 import { BsDynamicFormSelectComponent } from './dynamic-form-select.component';
 
 export const bsDynamicFormSelectType: DynamicFormInputType = {
@@ -17,6 +18,7 @@ export const bsDynamicFormSelectType: DynamicFormInputType = {
     CommonModule,
     ReactiveFormsModule,
     DynamicFormConfigModule.withInput(bsDynamicFormSelectType),
+    BsDynamicFormAddOnsModule,
   ],
   declarations: [
     BsDynamicFormSelectComponent,