+++ libs/bootstrap/src/lib/dynamic-forms.module.ts
@@ -3,7 +3,8 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormsModule, DynamicFormArrayModule, DynamicFormContainerModule, DynamicFormContentModule,
   DynamicFormControlModule, DynamicFormDictionaryModule, DynamicFormGroupModule, DynamicFormIdBuilder,
-  DynamicFormValidationModule, DYNAMIC_FORM_ID_BUILDER, DYNAMIC_FORM_LIBRARY, DYNAMIC_FORM_THEME } from '@dynamic-forms/core';
+  DynamicFormValidationModule, DYNAMIC_FORM_ID_BUILDER, DYNAMIC_FORM_LIBRARY, DYNAMIC_FORM_THEME,
+  DynamicFormTextModule } from '@dynamic-forms/core';
 import { BsDynamicFormActionModule } from './dynamic-form-action/dynamic-form-action.module';
 import { BsDynamicFormElementModule } from './dynamic-form-element/dynamic-form-element.module';
 import { BsDynamicFormFieldWrapperModule } from './dynamic-form-field/dynamic-form-field-wrapper.module';
@@ -22,6 +23,7 @@ import { bsDynamicFormLibrary } from './dynamic-form-library/dynamic-form-librar
     DynamicFormValidationModule.withValidation(),
     DynamicFormContainerModule,
     DynamicFormContentModule,
+    DynamicFormTextModule,
     BsDynamicFormElementModule,
     BsDynamicFormActionModule,
     BsDynamicFormInputModule,