+++ libs/core/src/lib/dynamic-form-action/dynamic-form-icon/dynamic-form-icon-base.spec.ts
@@ -16,8 +16,10 @@ import { DynamicFormIconTemplate } from './dynamic-form-icon-template';
   selector: 'dynamic-form-icon-test',
   template: `
     <button class="dynamic-form-icon"
-      [class.hidden]="template?.hidden" [ngClass]="template?.className"
-      [type]="template?.type || 'button'" [disabled]="template?.disabled"
+      [ngClass]="template.className"
+      [type]="template.type || 'button'"
+      [disabled]="template.disabled"
+      [hidden]="template.hidden"
       (click)="onClick($event)"
     >{{ template?.label }}</button>
   `,
@@ -82,12 +84,12 @@ describe('DynamicFormIconBase', () => {
     const formButtonDebugElement = fixture.debugElement.query(By.css('button.dynamic-form-icon'));
     const formButtonElement = formButtonDebugElement.nativeElement as HTMLButtonElement;
 
-    expect(formButtonElement.className).toBe('dynamic-form-icon');
+    expect(formButtonElement.hidden).toBeFalse();
 
     component.template.hidden = true;
     fixture.detectChanges();
 
-    expect(formButtonElement.className).toBe('dynamic-form-icon hidden');
+    expect(formButtonElement.hidden).toBeTrue();
   });
 
   it('sets class name of dynamic form icon', () => {