+++ libs/core/src/lib/dynamic-form-element/dynamic-form-text/dynamic-form-text-template.ts
@@ -1,0 +1,5 @@
+import { DynamicFormElementTemplate } from '../dynamic-form-element-template';
+
+export interface DynamicFormTextTemplate extends DynamicFormElementTemplate {
+  text: string;
+}