+++ libs/core/src/lib/dynamic-form-input/dynamic-form-input.ts
@@ -18,7 +18,14 @@ export interface DynamicFormInput<Value = any> {
 
 export type DynamicFormInputValue<FormInput> = FormInput extends DynamicFormInput<infer Value> ? Value : never;
 
-export type DynamicFormInputDefinition<FormInput extends DynamicFormInput> =
+export interface DynamicFormInputDefinition<
+  FormInput extends DynamicFormInput
+> extends DynamicFormControlDefinition<DynamicFormInputValue<FormInput>, FormInput> {
+  prefixAddOns?: undefined;
+  suffixAddOns?: undefined;
+}
+
+export type DynamicFormInputWithAddOnsDefinition<FormInput extends DynamicFormInput> =
   DynamicFormControlDefinition<DynamicFormInputValue<FormInput>, FormInput>;
 
 export class DynamicFormInputControl<FormInput extends DynamicFormInput>