+++ libs/material/src/lib/dynamic-form-input/dynamic-form-select/dynamic-form-select.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
-import { DynamicFormConfigModule, DynamicFormInputType } from '@dynamic-forms/core';
+import { DynamicFormConfigModule, DynamicFormElementModule, DynamicFormInputType } from '@dynamic-forms/core';
 import { matDynamicFormLibrary } from '../../dynamic-form-library/dynamic-form-library';
 import { MatDynamicFormSelectComponent } from './dynamic-form-select.component';
 
@@ -19,6 +19,7 @@ export const matDynamicFormSelectType: DynamicFormInputType = {
     ReactiveFormsModule,
     MatFormFieldModule,
     MatSelectModule,
+    DynamicFormElementModule,
     DynamicFormConfigModule.withInput(matDynamicFormSelectType),
   ],
   declarations: [